Update for 1.5.1
CH00138411 - Expose further LDAP Search Configuration options to the
Configuration File

LDAP Search Query is now saved to the Log file

Errors Creating / Updating Users as well as Assigning Roles are
returned to the Hornbill Log File for easier Diagnostics

Prevent "<invalid Value>" from being used as a value when an LDAP
Attribute cannot be found and instead an Error in the log is written.

Errors while create / Updating users are no loner output to the console
breaking the progress bar and instead an error count is output at
the end of the process and will refer the user to the Log File.